% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_server.R
\name{addLegend}
\alias{addLegend}
\title{Leaflet addLegend}
\usage{
addLegend(
  map,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  pal,
  values,
  na.label = "NA",
  bins = 7,
  colors,
  opacity = 0.5,
  labels = NULL,
  labFormat = leaflet::labelFormat(),
  title = NULL,
  className = "info legend",
  layerId = NULL,
  group = NULL,
  data = leaflet::getMapData(map),
  decreasing = FALSE
)
}
\arguments{
\item{map}{}

\item{position}{}

\item{pal}{}

\item{values}{}

\item{na.label}{}

\item{bins}{}

\item{colors}{}

\item{opacity}{}

\item{labels}{}

\item{labFormat}{}

\item{title}{}

\item{className}{}

\item{layerId}{}

\item{group}{}

\item{data}{}

\item{decreasing}{}
}
\value{
A leaflet legend
}
\description{
Leaflet addLegend
}
