% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoApp-package.R
\docType{package}
\name{GeoApp-package}
\alias{GeoApp}
\alias{GeoApp-package}
\title{GeoApp: Geography App}
\description{
A visualization tool for geography lessons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nenuial/GeoApp.git}
  \item Report bugs at \url{https://github.com/Nenuial/GeoApp.git/issues}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com}

}
\keyword{internal}
