% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_ui.R
\name{geoapp_ui_country_code}
\alias{geoapp_ui_country_code}
\title{Country codes for selector}
\usage{
geoapp_ui_country_code(code_type)
}
\arguments{
\item{code_type}{}
}
\value{
A named character vector
}
\description{
Country codes for selector
}
