% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_ui.R
\name{geoapp_ui_swiss_votes}
\alias{geoapp_ui_swiss_votes}
\title{Latest swiss votes and their ids}
\usage{
geoapp_ui_swiss_votes(votedate)
}
\value{
A named character vector
}
\description{
Latest swiss votes and their ids
}
