% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_ui.R
\name{geoapp_ui_swiss_votes_dates}
\alias{geoapp_ui_swiss_votes_dates}
\title{Swiss vote dates}
\usage{
geoapp_ui_swiss_votes_dates()
}
\value{
A named character vector
}
\description{
Swiss vote dates
}
