% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_ui.R
\name{geoapp_ui_wb_country_selector}
\alias{geoapp_ui_wb_country_selector}
\title{WB Country selector}
\usage{
geoapp_ui_wb_country_selector(id, label)
}
\arguments{
\item{id}{A string to use as input id}

\item{label}{A label to display for the selector}
}
\value{
An f7Select input object
}
\description{
WB Country selector
}
