% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_server.R
\name{get_country_name}
\alias{get_country_name}
\title{Get country name for specific code type}
\usage{
get_country_name(code, code_type)
}
\arguments{
\item{code}{Country code}

\item{code_type}{Code type in the countrycode package}
}
\value{

}
\description{
Get country name for specific code type
}
