% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_server.R
\name{get_wb_country_name}
\alias{get_wb_country_name}
\title{Get country name for WB ISO3 code}
\usage{
get_wb_country_name(code)
}
\arguments{
\item{iso3c}{A WB ISO3 code}
}
\value{
The country name
}
\description{
Get country name for WB ISO3 code
}
