% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demo_explorer.R
\name{mod_demo_explorer_ui}
\alias{mod_demo_explorer_ui}
\alias{mod_demo_explorer_server}
\title{mod_demo_explorer_ui and mod_demo_explorer_server}
\usage{
mod_demo_explorer_ui(id)

mod_demo_explorer_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
