% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hmd_demograph.R
\name{mod_hmd_demograph_ui}
\alias{mod_hmd_demograph_ui}
\alias{mod_hmd_demograph_server}
\title{mod_hmd_demograph_ui and mod_hmd_demograph_server}
\usage{
mod_hmd_demograph_ui(id)

mod_hmd_demograph_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
