% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_idb_pyramid.R
\name{mod_idb_pyramid_ui}
\alias{mod_idb_pyramid_ui}
\alias{mod_idb_pyramid_server}
\title{mod_idb_pyramid_ui and mod_idb_pyramid_server}
\usage{
mod_idb_pyramid_ui(id)

mod_idb_pyramid_server(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
