% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{tag_remove_attributes}
\alias{tag_remove_attributes}
\title{Remove a tag attribute}
\usage{
tag_remove_attributes(tag, ...)
}
\arguments{
\item{tag}{the tag}

\item{...}{the attributes to remove}
}
\value{
a new tag
}
\description{
Remove a tag attribute
}
\examples{
a <- shiny::tags$p(src = "plop", "pouet")
tag_remove_attributes(a, "src")
}
