% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoapp_utils_server.R
\name{add_legend}
\alias{add_legend}
\title{Leaflet add_legend}
\usage{
add_legend(
  map,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  pal,
  values,
  na_label = "NA",
  bins = 7,
  colors,
  opacity = 0.5,
  labels = NULL,
  lab_format = leaflet::labelFormat(),
  title = NULL,
  class_name = "info legend",
  layer_id = NULL,
  group = NULL,
  data = leaflet::getMapData(map),
  decreasing = FALSE
)
}
\arguments{
\item{map}{A leaflet map}

\item{position}{The position}

\item{pal}{The palettes}

\item{values}{Palettes values}

\item{bins}{Number of bins}

\item{colors}{The color}

\item{opacity}{The opacity}

\item{labels}{The labels}

\item{title}{The title}

\item{group}{The group}

\item{data}{The data}

\item{decreasing}{Decreasing?}

\item{na.label}{Label for na}

\item{labFormat}{Format for labels}

\item{className}{The class name}

\item{layerId}{The layer id}
}
\value{
A leaflet legend
}
\description{
Leaflet add_legend
}
